pyproject.toml
[tool.poetry]
name = "intui-pnid-langraph"
version = "0.1.0"
description = "A LangGraph project for parsing P&IDs into a knowledge graph."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langgraph = "^0.2.20"  # Ensure compatibility with langchain
langchain = "^0.3.9"  # Compatible version with updated pydantic
pydantic = ">=2.7.4,<3.0.0"  # Aligned with langchain requirements
ultralytics = "^8.3.39"
numpy = "^1.21.0"
Pillow = ">=9.2.0"  # Ensures compatibility with python-doctr
tavily-python = "^0.1.0"
pandas = "^1.3.0"
networkx = "^2.6.0"
azure-storage-blob = "^12.16.0"
python-doctr = "^0.10.0"  # Included for OCR functionality
torch = "^2.0.0"  # Required for Doctr OCR model support
tensorflow = "^2.13.0"  # Required for Doctr OCR
tf2onnx = "^1.16.1"
opencv-python-headless = "^4.10.0.84"  # Headless version for environments without GUI

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"

# Add these tools for additional development and testing support
pytest = "^7.1.2"
pytest-mock = "^3.10.0"
mypy = "^0.950"